Fix: "게시물 수정 및 삭제 오류 해결"
게시물 수정 및 삭제에서 로그인 member와 일정의 member가 일치할 때 예외를 던지는
문제가 발생하였습니다. 따라서, 조건문의 !를 제거하여 해결하였습니다.
